+++ src/MphRead/Formats/Formats.cs
@@ -278,92 +278,6 @@ namespace MphRead
             return new ColorRgba((byte)(color.Red * alpha), (byte)(color.Green * alpha), (byte)(color.Blue * alpha), 255);
         }
 
-        public void ExportImages()
-        {
-            string exportPath = Path.Combine(Paths.Export, Name);
-            Directory.CreateDirectory(exportPath);
-            foreach (Recolor recolor in Recolors)
-            {
-                string colorPath = Path.Combine(exportPath, recolor.Name);
-                Directory.CreateDirectory(colorPath);
-                var usedTextures = new HashSet<int>();
-                foreach (Material material in Materials.OrderBy(m => m.TextureId).ThenBy(m => m.PaletteId))
-                {
-                    if (material.TextureId == UInt16.MaxValue)
-                    {
-                        continue;
-                    }
-                    Texture texture = recolor.Textures[material.TextureId];
-                    IReadOnlyList<ColorRgba> pixels = recolor.GetPixels(material.TextureId, material.PaletteId);
-                    if (texture.Width == 0 || texture.Height == 0 || pixels.Count == 0)
-                    {
-                        continue;
-                    }
-                    Debug.Assert(texture.Width * texture.Height == pixels.Count);
-                    usedTextures.Add(material.TextureId);
-                    string filename = $"{material.TextureId}-{material.PaletteId}";
-                    SaveTexture(colorPath, filename, texture.Width, texture.Height, pixels);
-                }
-                if (usedTextures.Count != recolor.Textures.Count)
-                {
-                    string unusedPath = Path.Combine(colorPath, "unused");
-                    Directory.CreateDirectory(unusedPath);
-                    for (int t = 0; t < recolor.Textures.Count; t++)
-                    {
-                        if (usedTextures.Contains(t))
-                        {
-                            continue;
-                        }
-                        Texture texture = recolor.Textures[t];
-                        for (int p = 0; p < Palettes.Count; p++)
-                        {
-                            IReadOnlyList<TextureData> textureData = recolor.TextureData[t];
-                            IReadOnlyList<PaletteData> palette = recolor.PaletteData[p];
-                            if (textureData.Any(t => t.Data >= palette.Count))
-                            {
-                                continue;
-                            }
-                            IReadOnlyList<ColorRgba> pixels = recolor.GetPixels(t, p);
-                            string filename = $"{t}-{p}";
-                            SaveTexture(unusedPath, filename, texture.Width, texture.Height, pixels);
-                        }
-                    }
-                }
-            }
-        }
-
-        public void ExportPalettes()
-        {
-            string exportPath = Path.Combine(Paths.Export, Name);
-            Directory.CreateDirectory(exportPath);
-            foreach (Recolor recolor in Recolors)
-            {
-                string colorPath = Path.Combine(exportPath, recolor.Name);
-                Directory.CreateDirectory(colorPath);
-                string palettePath = Path.Combine(colorPath, "palettes");
-                Directory.CreateDirectory(palettePath);
-                for (int p = 0; p < recolor.Palettes.Count; p++)
-                {
-                    IReadOnlyList<ColorRgba> pixels = recolor.GetPalettePixels(p);
-                    string filename = $"p{p}";
-                    SaveTexture(palettePath, filename, 16, 16, pixels);
-                }
-            }
-        }
-
-        private void SaveTexture(string directory, string filename, ushort width, ushort height, IReadOnlyList<ColorRgba> pixels)
-        {
-            string imagePath = Path.Combine(directory, $"{filename}.png");
-            var bitmap = new Bitmap(width, height);
-            for (int p = 0; p < pixels.Count; p++)
-            {
-                ColorRgba pixel = pixels[p];
-                bitmap.SetPixel(p % width, p / width,
-                    Color.FromArgb(pixel.Alpha, pixel.Red, pixel.Green, pixel.Blue));
-            }
-            bitmap.Save(imagePath);
-        }
-
         private static void ThrowIfInvalidEnums(IEnumerable<RawMaterial> materials)
         {
             foreach (RawMaterial material in materials)
